% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-evidence-stack.R
\name{te_analytical_reference_data}
\alias{te_analytical_reference_data}
\title{Tabular representation of an analytical reference}
\usage{
te_analytical_reference_data(x)
}
\arguments{
\item{x}{te_analytical_reference.}
}
\value{
data.frame.
}
\description{
Tabular representation of an analytical reference
}
