% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-article-comparison.R
\name{te_article_case_set_boundaries}
\alias{te_article_case_set_boundaries}
\title{Existing boundary events for article cases}
\usage{
te_article_case_set_boundaries(x)
}
\arguments{
\item{x}{te_article_case_set.}
}
\value{
data.frame.
}
\description{
Existing boundary events for article cases
}
