% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-article-semantic-comparison.R
\name{te_article_case_set_transition_signature}
\alias{te_article_case_set_transition_signature}
\title{Transition signature for each article case}
\usage{
te_article_case_set_transition_signature(x)
}
\arguments{
\item{x}{te_article_case_set.}
}
\value{
data.frame.
}
\description{
Produces an ordered textual sequence of existing Phi transitions.
}
