% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-article-integration.R
\name{te_article_case_summary}
\alias{te_article_case_summary}
\title{Summarize an imported article case}
\usage{
te_article_case_summary(x)
}
\arguments{
\item{x}{te_article_case.}
}
\value{
One-row data.frame.
}
\description{
Summarize an imported article case
}
