% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-article-integration.R
\name{te_article_classify_object}
\alias{te_article_classify_object}
\title{Classify a first-article R object}
\usage{
te_article_classify_object(object)
}
\arguments{
\item{object}{R object from the article project.}
}
\value{
One-row data.frame.
}
\description{
Classify a first-article R object
}
