% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-article-integration.R
\name{te_article_experiment_registry}
\alias{te_article_experiment_registry}
\title{Registry of first-article experiment purposes}
\usage{
te_article_experiment_registry()
}
\value{
data.frame.
}
\description{
Registry of first-article experiment purposes
}
