% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-mlp-bridges.R
\name{te_bridge_ml_variable}
\alias{te_bridge_ml_variable}
\title{Bridge a real ML.DSNeuralRNAS learned variable}
\usage{
te_bridge_ml_variable(object)
}
\arguments{
\item{object}{ml_variable_aprendida.}
}
\value{
te_trace with exact parameter map.
}
\description{
Bridge a real ML.DSNeuralRNAS learned variable
}
