% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-evidence-stack.R
\name{te_evidence_analytical_contrast}
\alias{te_evidence_analytical_contrast}
\title{Compare a trace with an analytical reference when supported}
\usage{
te_evidence_analytical_contrast(x)
}
\arguments{
\item{x}{te_evidence_stack.}
}
\value{
data.frame.
}
\description{
Compare a trace with an analytical reference when supported
}
