% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-evidence-stack.R
\name{te_evidence_stack_levels}
\alias{te_evidence_stack_levels}
\title{Evidence-stack level table}
\usage{
te_evidence_stack_levels(x)
}
\arguments{
\item{x}{te_evidence_stack.}
}
\value{
data.frame.
}
\description{
Evidence-stack level table
}
