% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-formalspec.R
\name{te_formalspec_summary}
\alias{te_formalspec_summary}
\title{Summarize a complete FormalSpec analysis}
\usage{
te_formalspec_summary(x)
}
\arguments{
\item{x}{te_formalspec_analysis.}
}
\value{
One-row data.frame.
}
\description{
Summarize a complete FormalSpec analysis
}
