% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-spaces.R
\name{te_learning_space}
\alias{te_learning_space}
\title{Learning-dynamics 3D space}
\usage{
te_learning_space(trace)
}
\arguments{
\item{trace}{te_trace.}
}
\description{
x = normalized learning time,
y = log(loss / initial loss),
z = log(1 + gradient norm).
}
