% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-multirun-formalspec.R
\name{te_multirun_formalspec_summary}
\alias{te_multirun_formalspec_summary}
\title{FormalSpec summary across runs}
\usage{
te_multirun_formalspec_summary(x)
}
\arguments{
\item{x}{te_multirun_formalspec.}
}
\value{
data.frame.
}
\description{
FormalSpec summary across runs
}
