% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-multirun.R
\name{te_multirun_trace_data}
\alias{te_multirun_trace_data}
\title{Long canonical data for multiple runs}
\usage{
te_multirun_trace_data(x)
}
\arguments{
\item{x}{te_multirun.}
}
\value{
data.frame.
}
\description{
Normalized time tau is added for visual alignment only. Original iterations
are retained and no interpolation is performed.
}
