% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-trace.R
\name{te_normalized_time}
\alias{te_normalized_time}
\title{Normalized learning time}
\usage{
te_normalized_time(trace)
}
\arguments{
\item{trace}{te_trace or data.frame.}
}
\value{
Numeric vector in [0, 1].
}
\description{
Normalized learning time
}
