% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameter-map.R
\name{te_parameter_map_data}
\alias{te_parameter_map_data}
\title{Extract parameter-map data}
\usage{
te_parameter_map_data(x)
}
\arguments{
\item{x}{te_parameter_map.}
}
\description{
Extract parameter-map data
}
