% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26-visual-spec.R
\name{te_plotly_axis_spec}
\alias{te_plotly_axis_spec}
\title{Plotly axis specification}
\usage{
te_plotly_axis_spec(title, visual_spec, range = NULL)
}
\arguments{
\item{title}{Axis title.}

\item{visual_spec}{Result of te_visual_spec().}

\item{range}{Optional numeric range.}
}
\value{
list suitable for plotly layout.
}
\description{
Plotly axis specification
}
