% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-bridge-registry.R
\name{te_register_bridge}
\alias{te_register_bridge}
\title{Register a TraceExplorer bridge}
\usage{
te_register_bridge(
  source_class,
  extractor,
  source_package = NA_character_,
  description = "",
  normative = FALSE
)
}
\arguments{
\item{source_class}{Source S3 class.}

\item{extractor}{Function(object, ...) returning te_trace or canonical data.frame.}

\item{source_package}{Informational producer package.}

\item{description}{Human-readable description.}

\item{normative}{Whether the bridge is normative. Default FALSE.}
}
\description{
Register a TraceExplorer bridge
}
