% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-default-bridges.R
\name{te_register_default_bridges}
\alias{te_register_default_bridges}
\title{Register default ecosystem bridges}
\usage{
te_register_default_bridges()
}
\description{
Register default ecosystem bridges
}
