% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-article-comparison.R
\name{te_validate_article_case_set}
\alias{te_validate_article_case_set}
\title{Validate an article comparison set}
\usage{
te_validate_article_case_set(x, error = FALSE)
}
\arguments{
\item{x}{te_article_case_set.}

\item{error}{Stop on failure.}
}
\value{
Validation table.
}
\description{
Validate an article comparison set
}
