% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-evidence-stack.R
\name{te_validate_evidence_stack}
\alias{te_validate_evidence_stack}
\title{Validate an evidence stack}
\usage{
te_validate_evidence_stack(x, error = FALSE)
}
\arguments{
\item{x}{te_evidence_stack.}

\item{error}{Stop on failure.}
}
\value{
data.frame.
}
\description{
Validate an evidence stack
}
