% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-multirun.R
\name{te_validate_multirun}
\alias{te_validate_multirun}
\title{Validate a multi-run collection}
\usage{
te_validate_multirun(x, error = FALSE)
}
\arguments{
\item{x}{te_multirun.}

\item{error}{Stop on failure.}
}
\value{
Validation table.
}
\description{
Validate a multi-run collection
}
