% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-multirun-formalspec.R
\name{te_validate_multirun_formalspec}
\alias{te_validate_multirun_formalspec}
\title{Validate multi-run FormalSpec results}
\usage{
te_validate_multirun_formalspec(x, error = FALSE)
}
\arguments{
\item{x}{te_multirun_formalspec.}

\item{error}{Stop on failure.}
}
\value{
Validation table.
}
\description{
Validate multi-run FormalSpec results
}
