% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-contract.R
\name{te_validate_trace}
\alias{te_validate_trace}
\title{Validate a canonical trace data frame}
\usage{
te_validate_trace(x, error = TRUE)
}
\arguments{
\item{x}{data.frame.}

\item{error}{Whether to stop on invalid input.}
}
\value{
A validation table.
}
\description{
Validate a canonical trace data frame
}
