% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25-browser-visual-policy.R
\name{te_webgl_display_policy}
\alias{te_webgl_display_policy}
\title{Resolve browser visualization policy}
\usage{
te_webgl_display_policy(webgl_supported = NA)
}
\arguments{
\item{webgl_supported}{TRUE, FALSE, NULL or NA.}
}
\value{
One-row data.frame.
}
\description{
Resolve browser visualization policy
}
